refactor: move StatCard and PersonalStats to the designated folder
