+++ src/pages/Profile.jsx
@@ -10,7 +10,7 @@ import {
 
 // Sections
 import UserEditForm from "@/components/profile/userEditForm";
-const PersonalStats = lazy(() => import('@/components/stats/personalStats'));
+const PersonalStats = lazy(() => import('@/pages/statistics/personal/PersonalStats'));
 const MyDutiesCard = lazy(() => import('@/components/rota/myDuties'));
 const ModifyRotaMulti = lazy(() => import('@/components/profile/modifyRotaMulti'));
