+++ src/pages/statistics/personal/components/StatCard/StatCard.jsx
@@ -1,6 +1,8 @@
+import PropTypes from 'prop-types';
 import { Skeleton } from '@/components/ui/skeleton';
 import { OctagonX } from "lucide-react";
 
+/** Information card showing statistics and comparing it to the previous period */
 export default function StatCard ({ status, label, sublabel, value, previousValue }) {
   const showComparison = previousValue !== undefined && previousValue !== null;
   const isPositive = showComparison && value > previousValue;
@@ -17,7 +19,7 @@ export default function StatCard ({ status, label, sublabel, value, previousValu
         </div>
       )}
 
-      {status === "error" && (
+      {status === 'error' && (
         <>
           <div className="flex items-center gap-2 text-red-500 dark:text-red-400">
             <OctagonX className="h-4 w-4" />
@@ -46,3 +48,16 @@ export default function StatCard ({ status, label, sublabel, value, previousValu
     </div>
   );
 };
+
+StatCard.propTypes = {
+  /** API status */
+  status: PropTypes.oneOf(['pending', 'error', 'success']).isRequired,
+  /** Label for the period */
+  label: PropTypes.string.isRequired,
+  /** Sublabel for the period */
+  sublabel: PropTypes.string.isRequired,
+  /** Statistic value for the period */
+  value: PropTypes.number.isRequired,
+  /** Statistics value for the previous period */
+  previousValue: PropTypes.number
+}
